Moved utils to package root
