+++ pycox/evaluation/eval_surv.py
@@ -3,7 +3,8 @@ import warnings
 import numpy as np
 import pandas as pd
 from pycox.evaluation.concordance import concordance_td
-from pycox.evaluation import utils, ipcw
+from pycox.evaluation import ipcw
+from pycox import utils
 
 
 class EvalSurv:
